feat: Refactor NotionToolbarButton and introduce styling for improved user interaction
- Updated NotionToolbarButton component to enhance hover effects and tooltip functionality.
- Created a new NotionToolbarButton style file for better separation of concerns and improved styling management.
- Adjusted imports in related components to reflect the new structure, ensuring better organization and maintainability.
